Conflicts:
	Makefile.am
	cmd/arcstat/Makefile.am
	cmd/mount_zfs/mount_zfs.c
	cmd/zed/Makefile.am
	cmd/zpios/Makefile.am
	config/kernel.m4
	include/sys/dmu.h
	include/sys/zfs_context.h
	include/sys/zfs_ctldir.h
	include/sys/zfs_debug.h
	include/sys/zfs_vfsops.h
	include/sys/zpl.h
	lib/libavl/Makefile.am
	lib/libefi/Makefile.am
	lib/libnvpair/Makefile.am
	lib/libshare/Makefile.am
	lib/libspl/Makefile.am
	lib/libspl/include/sys/Makefile.am
	lib/libspl/include/sys/mount.h
	lib/libunicode/Makefile.am
	lib/libuutil/Makefile.am
	lib/libzfs/Makefile.am
	lib/libzfs/libzfs_dataset.c
	lib/libzfs_core/Makefile.am
	lib/libzpool/Makefile.am
	lib/libzpool/kernel.c
	module/Makefile.in
	module/zfs/Makefile.in
	module/zfs/arc.c
	module/zfs/dbuf.c
	module/zfs/dmu.c
	module/zfs/dsl_pool.c
	module/zfs/vdev_disk.c
	module/zfs/zfs_ctldir.c
	module/zfs/zfs_debug.c
	module/zfs/zfs_ioctl.c
	module/zfs/zfs_vfsops.c
	module/zfs/zfs_vnops.c
	module/zfs/zfs_znode.c
	module/zfs/zpl_super.c
	module/zfs/zvol.c

+++ module/zfs/dbuf.c
@@ -615,7 +615,7 @@ dbuf_init(void)
 	 * By default, the table will take up
 	 * totalmem * sizeof(void*) / 8K (1MB per GB with 8-byte pointers).
 	 */
-	while (hsize * zfs_arc_average_blocksize < physmem * PAGESIZE)
+		while (hsize * zfs_arc_average_blocksize <(uint64_t)physmem * PAGESIZE)
 		hsize <<= 1;
 
 retry:
